Removes struct mock generation
This change removes the ability to generate mocks for structs.

Structs are value types, and therefore mocks are less useful.
Instead of mocking a struct, tests can now use actual struct instances.

The file pattern option is also removed as it is no longer needed.
Options for annotations are no longer used.